moved tests, got the happy path running
