+++ src/applications/health-care-questionnaire/questionnaire/tests/e2e/bugs/questionnaire.unauth.loading.bug.cypress.spec.js
@@ -1,4 +1,4 @@
-import featureToggles from './fixtures/mocks/feature-toggles.enabled.json';
+import featureToggles from '../fixtures/mocks/feature-toggles.enabled.json';
 
 describe('health care questionnaire -- un-authenticated', () => {
   beforeEach(() => {
