+++ src/applications/health-care-questionnaire/shared/tests/todo.md
@@ -1,0 +1,4 @@
+# Testing todo
+
+- [ ] fix form tester stuff
+- [ ] create a full end to end test (/Users/markdewey/va/vets-website/src/applications/health-care-questionnaire/list/tests/e2e/04.questionnaire.happy.path.cypress.spec.js)
